creating file structure and navigation
