+++ lib/screens/login/email_otp.dart
@@ -91,7 +91,9 @@ class EmailOtp extends StatelessWidget {
         Padding(
           padding: const EdgeInsets.symmetric(horizontal: 80.0),
           child: ElevatedButton(
-            onPressed: () {},
+            onPressed: () {
+              Navigator.pushReplacementNamed(context, '/phone_otp');
+            },
             child: Text(
               "Next",
               style: TextStyle(fontSize: 20,
