+++ lib/screens/login/password_changed.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-class ConfirmedPage extends StatelessWidget {
+class PasswordChanged extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return SafeArea(
@@ -61,7 +61,10 @@ class ConfirmedPage extends StatelessWidget {
         Padding(
           padding: const EdgeInsets.symmetric(horizontal: 80.0),
           child: ElevatedButton(
-            onPressed: () {},
+            onPressed: () {
+              Navigator.pushReplacementNamed(context, '/login');
+
+            },
             child: Text(
               "Login",
               style: TextStyle(fontSize: 20,
