+++ lib/screens/splash_screen.dart
@@ -1,3 +1,5 @@
+// start_page.dart
+
 import 'package:flutter/material.dart';
 
 class StartPage extends StatelessWidget {
@@ -40,22 +42,24 @@ class StartPage extends StatelessWidget {
       crossAxisAlignment: CrossAxisAlignment.center,
       children: [
         Image.asset(
-          'asset/images/logo.png', // Make sure to add your image to the assets folder and update the path
+          'asset/images/logo-box.jpg', // Make sure to add your image to the assets folder and update the path
           height: 100, // Adjust the height as needed
         ),
         SizedBox(height: 3),
         Text(
           "Confee",
-          style: TextStyle(fontSize: 50,
-              color: Colors.white,
-              fontWeight: FontWeight.bold
+          style: TextStyle(
+            fontSize: 50,
+            color: Colors.white,
+            fontWeight: FontWeight.bold,
           ),
         ),
         SizedBox(height: 1),
         Text(
           "ICTer Conference",
-          style: TextStyle(fontSize: 30,
-              color: Colors.white
+          style: TextStyle(
+            fontSize: 30,
+            color: Colors.white,
           ),
         ),
       ],
@@ -76,11 +80,14 @@ class StartPage extends StatelessWidget {
         Padding(
           padding: const EdgeInsets.symmetric(horizontal: 80.0),
           child: ElevatedButton(
-            onPressed: () {},
+            onPressed: () {
+              Navigator.pushReplacementNamed(context, '/login'); // Navigate to login screen
+            },
             child: Text(
               "Get Start",
-              style: TextStyle(fontSize: 20,
-                  color: Colors.white
+              style: TextStyle(
+                fontSize: 20,
+                color: Colors.white,
               ), // Text color is white
             ),
             style: ElevatedButton.styleFrom(
