+++ lib/screens/sponsers/description_page.dart
@@ -72,8 +72,11 @@
 // ignore_for_file: prefer_const_constructors
 
 import 'package:flutter/material.dart';
-import 'package:youtube_1/screens/sponsers/profile_view.dart'; // Import the ProfileViewPage
-import 'package:youtube_1/pages/participant/innovation/custom_bottom_navigation_bar.dart'; // Import CustomBottomNavigationBar
+// import 'package:youtube_1/screens/sponsers/profile_view.dart'; // Import the ProfileViewPage
+// import 'package:youtube_1/pages/participant/innovation/custom_bottom_navigation_bar.dart';
+
+import 'custom_bottom_navigation_bar.dart';
+import 'profile_view.dart'; // Import CustomBottomNavigationBar
 
 class DescriptionPage extends StatefulWidget {
   const DescriptionPage({Key? key}) : super(key: key);
